Handle failed workers on time list generation
Previously the time list generation assumed that all the workers worked
and would block until all the data was collected from the queue. But if
the workers failed for whatever reason things would get stuck on the
blocking get() call. This commit fixes this by adding the timeout to
get() and then reporting the number of failed clients on the output
before moving on.